+++ .babelrc
@@ -1,0 +1,9 @@
+{
+  "presets": ["@babel/env"],
+  "plugins": [
+    ["@babel/plugin-proposal-decorators", {"legacy": true}],
+    ["@babel/plugin-proposal-class-properties", {"loose": true}],
+    ["@babel/plugin-proposal-optional-chaining", {"loose": true}],
+    ["@babel/plugin-proposal-nullish-coalescing-operator", {"loose": true}]
+  ]
+}
