+++ ecosystem.config.js
@@ -1,9 +1,9 @@
 module.exports = {
-  apps : [{
-    name: "infobot",
-    script: './index.js',
+  apps: [{
+    name: 'infobot',
+    script: './dist/index.js',
     watch: '.',
-    instances : "max",
-    exec_mode : "cluster"
-  }],
-};
+    instances: 'max',
+    exec_mode: 'cluster'
+  }]
+}
