+++ package.json
@@ -6,9 +6,12 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "pm2-runtime start ecosystem.config.js --env production",
-    "start:dev": "nodemon index.js",
-    "start:debug": "nodemon --inspect index.js",
-    "fix": "eslint --fix ./src index.js"
+    "start:dev": "nodemon index.js --exec babel-node",
+    "start:debug": "nodemon --inspect index.js --exec babel-node",
+    "build": "babel src --out-dir dist",
+    "fix": "eslint --fix ./src index.js",
+    "clean": "rimraf ./dist",
+    "typeorm": "babel-node ./node_modules/typeorm/cli.js"
   },
   "author": "",
   "license": "ISC",
@@ -20,12 +23,21 @@
     "pm2": "^4.4.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.6",
+    "@babel/node": "^7.8.7",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-decorators": "^7.8.3",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
+    "@babel/plugin-proposal-optional-chaining": "^7.9.0",
+    "@babel/preset-env": "^7.9.6",
     "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.1",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
-    "nodemon": "^2.0.3"
+    "nodemon": "^2.0.3",
+    "rimraf": "^3.0.2"
   }
 }
